dom/filesystem/tests/test_basic.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Directory API</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>

<body>
<input id="fileList" type="file"></input>
<script type="application/javascript;version=1.7">

var directory;

function create_fileList() {
  var url = SimpleTest.getTestFileURL("script_fileList.js");
  var script = SpecialPowers.loadChromeScript(url);

  function onOpened(message) {
    var fileList = document.getElementById('fileList');
    SpecialPowers.wrap(fileList).mozSetDirectory(message.dir);

    // Just a simple test
    is(fileList.files.length, 1, "Filelist has 1 element");
    ok(fileList.files[0] instanceof Directory, "We have a directory.");

    directory = fileList.files[0];

    script.destroy();
    next();
  }

  script.addMessageListener("dir.opened", onOpened);
  script.sendAsyncMessage("dir.open");
}

function test_basic() {
  ok(directory, "Directory exists.");
  ok(directory instanceof Directory, "We have a directory.");
  is(directory.name, '/', "directory.name must be '/'");
  is(directory.path, '/', "directory.path must be '/'");
  next();
}

function checkSubDir(dir) {
  return dir.getFilesAndDirectories().then(
    function(data) {
      for (var i = 0; i < data.length; ++i) {
        ok (data[i] instanceof File || data[i] instanceof Directory, "Just Files or Directories");
        if (data[i] instanceof Directory) {
          isnot(data[i].name, '/', "Subdirectory should be called with the leafname");
          isnot(data[i].path, '/', "Subdirectory path should be called with the leafname");
        }
      }
    }
  );
}

function getFilesAndDirectories() {
  directory.getFilesAndDirectories().then(
    function(data) {
      ok(data.length, "We should have some data.");
      var promises = [];
      for (var i = 0; i < data.length; ++i) {
        ok (data[i] instanceof File || data[i] instanceof Directory, "Just Files or Directories");
        if (data[i] instanceof Directory) {
          isnot(data[i].name, '/', "Subdirectory should be called with the leafname");
          isnot(data[i].path, '/', "Subdirectory path should be called with the leafname");
          promises.push(checkSubDir(data[i]));
        }
      }

      return Promise.all(promises);
    },
    function() {
      ok(false, "Something when wrong");
    }
  ).then(next);
}

var tests = [
  create_fileList,

  test_basic,

  getFilesAndDirectories,
];

function next() {
  if (!tests.length) {
    SimpleTest.finish();
    return;
  }

  var test = tests.shift();
  test();
}

SimpleTest.waitForExplicitFinish();
next();
</script>
</body>
</html>
